refactor(Further Cell Analysis): Refactor
- Refactored further cell analysis into its own component
