+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2022-02-14T09:06:25.081Z\n"
-"PO-Revision-Date: 2022-02-14T09:06:25.081Z\n"
+"POT-Creation-Date: 2022-02-14T09:25:42.600Z\n"
+"PO-Revision-Date: 2022-02-14T09:25:42.600Z\n"
 
 msgid "Next"
 msgstr "Next"
@@ -1331,8 +1331,8 @@ msgstr "Dictionary"
 msgid "Further Analysis"
 msgstr "Further Analysis"
 
-msgid "OrgUnit Analysis "
-msgstr "OrgUnit Analysis "
+msgid "OrgUnit Analysis"
+msgstr "OrgUnit Analysis"
 
 msgid "Trend Analysis "
 msgstr "Trend Analysis "
