+++ src/modules/Main/Components/ScorecardView/Components/ScorecardTable/Components/TableDataContainer/Components/FurtherAnalysisMenu/index.js
@@ -1,0 +1,139 @@
+import i18n from "@dhis2/d2-i18n";
+import {
+    IconDimensionOrgUnit16,
+    IconVisualizationColumnStacked24,
+    IconVisualizationLine24,
+    Menu,
+    MenuItem,
+    Popover
+} from "@dhis2/ui";
+import {Period} from "@iapps/period-utilities";
+import React, {useState} from "react";
+import {useSetRecoilState} from "recoil";
+import TableCellAnalysis from "../../../TableCellAnalysis";
+import {orgUnitSelectorOptionOnCell} from "../../../TableCellAnalysis/state/orgUnit";
+import {cellPeriodOptionSelector} from "../../../TableCellAnalysis/state/period";
+import PropTypes from 'prop-types'
+
+export default function FurtherAnalysisMenu({
+                                                orgUnit,
+                                                period,
+                                                dataSources,
+                                                stateActionRef,
+                                                setStateActionRef,
+                                                analysisOpen,
+                                                setAnalysisOpen
+                                            }) {
+    const [showSubMenu, setShowSubMenu] = useState(false)
+    const setPeriodOptionValueStates = useSetRecoilState(cellPeriodOptionSelector);
+    const setOrgUnitOptionValueStates = useSetRecoilState(orgUnitSelectorOptionOnCell);
+
+    return (
+        <>
+            <Popover
+                onClickOutside={() => setStateActionRef(undefined)}
+                placement="bottom-start"
+                reference={stateActionRef}
+            >
+                <Menu>
+                    <MenuItem
+                        onClick={() => {
+                            setStateActionRef(undefined);
+                            setAnalysisOpen(true);
+                        }}
+                        label={i18n.t("Further Analysis")}
+                        icon={<IconVisualizationColumnStacked24/>}
+                    />
+                    <MenuItem
+                        onClick={() => {
+                            setOrgUnitOptionValueStates(true);
+                            setStateActionRef(undefined);
+                            setAnalysisOpen(true);
+                        }}
+                        label={i18n.t("OrgUnit Analysis")}
+                        icon={<IconDimensionOrgUnit16/>}
+                    />
+                    <MenuItem
+                        onClick={() => {
+                            setStateActionRef(undefined);
+                        }}
+                        showSubMenu={showSubMenu}
+                        toggleSubMenu={() => {
+                            setShowSubMenu(prevState => !prevState);
+                        }}
+                        label={i18n.t("Trend Analysis ")}
+                        icon={<IconVisualizationLine24/>}
+                    >
+                        <MenuItem
+                            onClick={() => {
+                                setStateActionRef(undefined);
+                                setPeriodOptionValueStates([new Period().setPreferences({allowFuturePeriods: true}).getById("LAST_3_MONTHS")])
+                                setAnalysisOpen(true);
+                            }}
+                            label={i18n.t("Last 3 Months")}
+                            icon={<IconVisualizationColumnStacked24/>}
+                        />
+                        <MenuItem
+                            onClick={() => {
+                                setStateActionRef(undefined);
+                                setPeriodOptionValueStates([new Period().setPreferences({allowFuturePeriods: true}).getById("LAST_6_MONTHS")])
+                                setAnalysisOpen(true);
+                            }}
+                            label={i18n.t("Last 6 Months")}
+                            icon={<IconVisualizationColumnStacked24/>}
+                        />
+                        <MenuItem
+                            onClick={() => {
+                                setStateActionRef(undefined);
+                                setPeriodOptionValueStates([new Period().setPreferences({allowFuturePeriods: true}).getById("LAST_12_MONTHS")])
+                                setAnalysisOpen(true);
+                            }}
+                            label={i18n.t("Last 12 Months")}
+                            icon={<IconVisualizationColumnStacked24/>}
+                        />
+                        <MenuItem
+                            onClick={() => {
+                                setStateActionRef(undefined);
+                                setPeriodOptionValueStates([new Period().setPreferences({allowFuturePeriods: true}).getById("LAST_4_QUARTERS")])
+                                setAnalysisOpen(true);
+                            }}
+                            label={i18n.t("Last 4 Quarters")}
+                            icon={<IconVisualizationColumnStacked24/>}
+                        />
+                        <MenuItem
+                            onClick={() => {
+                                setStateActionRef(undefined);
+                                setPeriodOptionValueStates([new Period().setPreferences({allowFuturePeriods: true}).getById("LAST_5_YEARS")])
+                                setAnalysisOpen(true);
+                            }}
+                            label={i18n.t("Last 5 Years")}
+                            icon={<IconVisualizationColumnStacked24/>}
+                        />
+                    </MenuItem>
+
+                </Menu>
+            </Popover>
+            {analysisOpen && (
+                <TableCellAnalysis
+                    orgUnit={orgUnit}
+                    period={period}
+                    dataHolder={{dataSources}}
+                    onClose={() => {
+                        setAnalysisOpen(false);
+                    }}
+                />
+            )}
+        </>
+    )
+}
+
+FurtherAnalysisMenu.propTypes = {
+    dataSources: PropTypes.array.isRequired,
+    orgUnit: PropTypes.object.isRequired,
+    period: PropTypes.object.isRequired,
+    analysisOpen: PropTypes.bool,
+    setAnalysisOpen: PropTypes.func,
+    setStateActionRef: PropTypes.func,
+    stateActionRef: PropTypes.object,
+};
+
